# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PDPS Backend API",
        "description": "API documentation for PDPS (Pradeshiya Sabha Development Project System) Backend",
        "contact": {
            "email": "admin@pdps.lk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/member": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get all members",
                "description": "Retrieve a list of all members",
                "operationId": "1937cebfb899e61404b10ccac91c2ff0",
                "responses": {
                    "200": {
                        "description": "Members retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Create a new member",
                "description": "Add a new member to the system",
                "operationId": "caf907ab40abe78111573fac76a4991c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nameEn",
                                    "nameSi",
                                    "nameTa"
                                ],
                                "properties": {
                                    "nameEn": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "nameSi": {
                                        "type": "string",
                                        "example": "ජෝන් ඩෝ"
                                    },
                                    "nameTa": {
                                        "type": "string",
                                        "example": "ஜான் டோ"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "mobileNo": {
                                        "type": "string",
                                        "example": "0771234567"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Member created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/member/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get a specific member",
                "description": "Retrieve details of a specific member",
                "operationId": "aefb8fa386df11a8f0ac4ffa33e531d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Member not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Update a member",
                "description": "Update an existing member",
                "operationId": "8714fc0defe618942b69205c32faa370",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nameEn",
                                    "nameSi",
                                    "nameTa",
                                    "email",
                                    "tel",
                                    "division",
                                    "party",
                                    "position"
                                ],
                                "properties": {
                                    "nameEn": {
                                        "type": "string",
                                        "example": "Updated John Doe"
                                    },
                                    "nameSi": {
                                        "type": "string",
                                        "example": "යාවත්කාලීන ජෝන් ඩෝ"
                                    },
                                    "nameTa": {
                                        "type": "string",
                                        "example": "புதுப்பிக்கப்பட்ட ஜான் டோ"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.updated@example.com"
                                    },
                                    "tel": {
                                        "type": "string",
                                        "example": "0771234567"
                                    },
                                    "division": {
                                        "type": "string",
                                        "example": "Division 1"
                                    },
                                    "party": {
                                        "type": "string",
                                        "example": "Party A"
                                    },
                                    "position": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Position 1",
                                            "Position 2"
                                        ]
                                    },
                                    "img": {
                                        "description": "Member image",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Member updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Member updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Member not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your API token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}